requirements.txt
google-generativeai==0.3.1
streamlit==1.29.0
joblib==1.3.2